Refactor App and HomeScreen to centralize database setup; move initialization logic to App component and remove redundant calls from HomeScreen
